Adversy2/requirements.txt
requests
python-dotenv
rich
vaderSentiment
beautifulsoup4
gradio
openai
reportlab
